Fix ARM Skia for non-NEON builds
Optimizations sometimes need code review. In this case, it is change
with the ID of Ic5cc63b3df86eb7e8ea0bbb3b7bddfc73db20025 that is to
blame for build breakage in cases where either __ARM_HAVE_NEON is not
defined or SK_CPU_LENDIAN is not defined. This commit patches up the
relevant part of the codebase so that the builder carries on. Some
minor indenting cleanup on the relevant code is also included.

Change-Id: I0324c4f6db0b024c0b6a218b563aa2e82344243c